+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # next (unreleased)
 
 * Make CSS style prettier (Jendrik Seipp).
-* Remove `--toc-only` option (Jendrik Seipp).
+* Remove `--toc-only` and `--toc-level` options (Jendrik Seipp).
 
 # 3.6 (2020-02-05)